requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
construct==2.8.8
crccheck==1.3.1
cryptography==45.0.5
ECPy==1.2.5
frozenlist==1.7.0
humanfriendly==10.0
idna==3.10
multidict==6.6.2
propcache==0.3.2
protobuf==4.25.8
pycparser==2.22
pycryptodome==3.23.0
pycryptodomex==3.23.0
pymp4==1.4.0
pyplayready==0.6.0
pyreadline3==3.5.4
pywidevine==1.8.0
PyYAML==6.0.2
requests==2.32.4
typing_extensions==4.14.0
Unidecode==1.4.0
urllib3==2.5.0
xmltodict==0.14.2
yarl==1.20.1
